moved all prompts and regex dep to cms-cli
