+++ packages/cms-cli/lib/regex.js
